requirements.txt
google-generativeai
supabase
python-telegram-bot==20.7
python-dotenv
